requirements.txt
Django
psycopg2-binary
python-telegram-bot==13.7
pyyaml
requests